CoastCapitalFinance/requirements.txt
# Web framework
flask==3.0.3
gunicorn==22.0.0

# Database
SQLAlchemy==2.0.36
pymysql==1.1.1

# Stock data sources
yfinance==0.2.48
pandas==2.2.3
numpy==2.0.2

# Technical analysis
ta==0.11.0

# Machine learning / forecasting
scikit-learn==1.5.2
lightgbm==4.5.0
xgboost==2.1.3
catboost==1.2.7
optuna==3.6.1
scipy==1.14.1

# LLM / AI
anthropic==0.40.0

# News & external data
requests==2.32.3
feedparser==6.0.11

# Data processing
joblib==1.4.2

# Config / env
pydantic==2.10.3
pydantic-settings==2.6.1

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0

# Utilities
tenacity==9.0.0
prometheus_client==0.21.0
